+++ Examples/GLModel Demo/GLModelExample/ViewController.m
@@ -101,7 +101,7 @@
             
             //set model
             modelView.texture = nil;
-            modelView.blendColor = [UIColor colorWithRed:0.0f green:0.0f blue:1.0f alpha:0.5f];
+            modelView.blendColor = nil;
             modelView.model = [GLModel modelWithContentsOfFile:@"cube.obj"];
             
             //set default transform